Finished refactoring code
