+++ pipexb/Makefile
@@ -13,7 +13,7 @@ CFLAGS := -Wall -Wextra -Werror
 # **************************************************************************** #
 # Source
 # **************************************************************************** #
-SRC_FILES			:=	main.c	utils.c
+SRC_FILES			:=	main.c	utils.c	utils2.c
 OBJS				=	$(SRC_FILES:.c=.o)
 
 # **************************************************************************** #
