+++ pipexb/utils2.h
@@ -1,0 +1,10 @@
+#ifndef UTILS2_H
+# define UTILS2_H
+
+# include "utils.h"
+
+int		p3child(char *envp[], struct s_pipex pipexstruct);
+
+int		checkforheredoc(char *argv[], struct s_pipex *pipexstruct);
+
+#endif
